+++ mobile/android/android-components/docs/changelog.md
@@ -12,8 +12,18 @@ permalink: /changelog/
 * [Gecko](https://github.com/mozilla-mobile/android-components/blob/master/buildSrc/src/main/java/Gecko.kt)
 * [Configuration](https://github.com/mozilla-mobile/android-components/blob/master/buildSrc/src/main/java/Config.kt)
 
+* **browser-engine-gecko-nightly**  
+  * Now supports window requests. A new tab will be opened for `target="_blank"` links and `window.open` calls.
+
 * **feature-app-links**
   * Fixed [#3944](https://github.com/mozilla-mobile/android-components/issues/3944) causing third-party apps being opened when links with a `javascript` scheme are clicked.
+
+* **feature-session**
+  * ⚠️ **This is a breaking change**:
+  * The `WindowFeature` no longer needs and engine can now be created using just: 
+  ```kotlin
+     val windowFeature = WindowFeature(components.sessionManager)
+  ```  
 
 # 6.0.0
 
@@ -31,6 +41,7 @@ permalink: /changelog/
 
 * **browser-engine-gecko-nightly**
   * The component now handles situations where the Android system kills the content process (without killing the main app process) in order to reclaim resources. In those situations the component will automatically recover and restore the last known state of those sessions.
+  * Now supports window requests. A new tab will be opened for `target="_blank"` links and `window.open` calls.
 
 * **service-location**
   * 🆕 A new component for accessing Mozilla's and other location services.